# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Monthly payout calculation for couriers with tiered commission.",
        "title": "Payout Calculation Service",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/couriers/{id}/payouts": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "couriers"
                ],
                "summary": "Courier payout history",
                "parameters": [
                    {
                        "type": "string",
                        "description": "courier id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "calculated | paid | cancelled",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "YYYY-MM",
                        "name": "period_from",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "YYYY-MM",
                        "name": "period_to",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "period | net_amount | created_at",
                        "name": "sort_by",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "asc | desc",
                        "name": "sort_dir",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "page size, max 100",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "offset",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_interfaces_http_handler.PayoutListResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_interfaces_http_handler.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/internal_interfaces_http_handler.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/internal_interfaces_http_handler.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/payouts": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "payouts"
                ],
                "summary": "List all payouts",
                "parameters": [
                    {
                        "type": "string",
                        "description": "courier id",
                        "name": "courier_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "calculated | paid | cancelled",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "YYYY-MM",
                        "name": "period_from",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "YYYY-MM",
                        "name": "period_to",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "period | net_amount | created_at",
                        "name": "sort_by",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "asc | desc",
                        "name": "sort_dir",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "page size, max 100",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "offset",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_interfaces_http_handler.PayoutListResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_interfaces_http_handler.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/internal_interfaces_http_handler.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/internal_interfaces_http_handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/payouts/calculate": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Idempotent. A second call for the same courier and month returns 409 with the existing payout.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "payouts"
                ],
                "summary": "Calculate payout for courier + month",
                "parameters": [
                    {
                        "description": "courier id and period YYYY-MM",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_interfaces_http_handler.calculatePayoutRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/internal_interfaces_http_handler.PayoutResponse"
                        }
                    },
                    "400": {
                        "description": "bad or future period",
                        "schema": {
                            "$ref": "#/definitions/internal_interfaces_http_handler.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/internal_interfaces_http_handler.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/internal_interfaces_http_handler.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "courier not found",
                        "schema": {
                            "$ref": "#/definitions/internal_interfaces_http_handler.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "already calculated",
                        "schema": {
                            "$ref": "#/definitions/internal_interfaces_http_handler.PayoutConflictResponse"
                        }
                    }
                }
            }
        },
        "/payouts/jobs/monthly": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Calculates payouts for all active couriers for one month. Default is the previous month. Same advisory lock as the cron run, a second parallel call returns skipped=true.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "payouts"
                ],
                "summary": "Run monthly payout job",
                "parameters": [
                    {
                        "description": "optional period",
                        "name": "body",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/internal_interfaces_http_handler.runJobRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_ruziba3vich_payout-calculation-service_internal_application_payout.JobResult"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_interfaces_http_handler.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/internal_interfaces_http_handler.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/internal_interfaces_http_handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/payouts/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "payouts"
                ],
                "summary": "Get payout with adjustments",
                "parameters": [
                    {
                        "type": "string",
                        "description": "payout id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_interfaces_http_handler.PayoutWithAdjustmentsResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/internal_interfaces_http_handler.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/internal_interfaces_http_handler.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/internal_interfaces_http_handler.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "github_com_ruziba3vich_payout-calculation-service_internal_application_payout.JobResult": {
            "type": "object",
            "properties": {
                "created": {
                    "type": "integer"
                },
                "existing": {
                    "type": "integer"
                },
                "failed": {
                    "type": "integer"
                },
                "period": {
                    "type": "string"
                },
                "reason": {
                    "type": "string"
                },
                "skipped": {
                    "type": "boolean"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
        "github_com_ruziba3vich_payout-calculation-service_internal_domain_payout.AdjustmentType": {
            "type": "string",
            "enum": [
                "order_returned",
                "order_cancelled",
                "order_added",
                "manual"
            ],
            "x-enum-varnames": [
                "AdjustmentOrderReturned",
                "AdjustmentOrderCancelled",
                "AdjustmentOrderAdded",
                "AdjustmentManual"
            ]
        },
        "github_com_ruziba3vich_payout-calculation-service_internal_domain_payout.Status": {
            "type": "string",
            "enum": [
                "calculated",
                "paid",
                "cancelled"
            ],
            "x-enum-varnames": [
                "StatusCalculated",
                "StatusPaid",
                "StatusCancelled"
            ]
        },
        "internal_interfaces_http_handler.AdjustmentResponse": {
            "type": "object",
            "properties": {
                "commission_delta": {
                    "type": "number"
                },
                "created_at": {
                    "type": "string"
                },
                "gross_delta": {
                    "type": "number"
                },
                "id": {
                    "type": "string"
                },
                "net_delta": {
                    "type": "number"
                },
                "order_id": {
                    "type": "string"
                },
                "payout_id": {
                    "type": "string"
                },
                "reason": {
                    "type": "string"
                },
                "type": {
                    "$ref": "#/definitions/github_com_ruziba3vich_payout-calculation-service_internal_domain_payout.AdjustmentType"
                }
            }
        },
...
        "internal_interfaces_http_handler.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
...
        "internal_interfaces_http_handler.PayoutConflictResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "payout": {
                    "$ref": "#/definitions/internal_interfaces_http_handler.PayoutResponse"
                }
            }
        },
        "internal_interfaces_http_handler.PayoutListResponse": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/internal_interfaces_http_handler.PayoutResponse"
                    }
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "internal_interfaces_http_handler.PayoutResponse": {
            "type": "object",
            "properties": {
                "calculated_at": {
                    "type": "string"
                },
                "commission_amount": {
                    "type": "number"
                },
                "commission_rate": {
                    "type": "number"
                },
                "courier_id": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "delivered_count": {
                    "type": "integer"
                },
                "gross_amount": {
                    "type": "number"
                },
                "id": {
                    "type": "string"
                },
                "net_amount": {
                    "type": "number"
                },
                "paid_at": {
                    "type": "string"
                },
                "period": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/github_com_ruziba3vich_payout-calculation-service_internal_domain_payout.Status"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "internal_interfaces_http_handler.PayoutWithAdjustmentsResponse": {
            "type": "object",
            "properties": {
                "adjustments": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/internal_interfaces_http_handler.AdjustmentResponse"
                    }
                },
                "calculated_at": {
                    "type": "string"
                },
                "commission_amount": {
                    "type": "number"
                },
                "commission_rate": {
                    "type": "number"
                },
                "courier_id": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "delivered_count": {
                    "type": "integer"
                },
                "gross_amount": {
                    "type": "number"
                },
                "id": {
                    "type": "string"
                },
                "net_amount": {
                    "type": "number"
                },
                "paid_at": {
                    "type": "string"
                },
                "period": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/github_com_ruziba3vich_payout-calculation-service_internal_domain_payout.Status"
                },
                "total_amount": {
                    "description": "TotalAmount is net_amount plus all adjustment net deltas.",
                    "type": "number"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "internal_interfaces_http_handler.calculatePayoutRequest": {
            "type": "object",
            "required": [
                "courier_id",
                "period"
            ],
            "properties": {
                "courier_id": {
                    "type": "string"
                },
                "period": {
                    "type": "string"
                }
            }
        },
...
        "internal_interfaces_http_handler.runJobRequest": {
            "type": "object",
            "properties": {
                "period": {
                    "description": "optional YYYY-MM, default previous month",
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer \" followed by the token from /auth/admin/login or /auth/courier/login.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}